fixed relationship between timeout & maxDepth
If the console prints 'The Computer reached his maximum computing time. He is now forced to play', the AI will not have searched all branches of the search tree evenly and thus be likely to make a flawed move. As a consequence, the combination of the two public variables timeout and maxDepth should always be tested for this condition, should one expect an optimal game by the AI.